+++ install/create_console_objects.sql
@@ -48,12 +48,16 @@ begin
     dbms_output.put_line('- Table CONSOLE_CONF not found, run creation command');
     execute immediate q'{
       create table console_conf (
-        conf_id           varchar2 (16 byte)  not null  ,
-        conf_by           varchar2 (64 byte)            ,
-        conf_sysdate      date                not null  ,
-        level_id          number   ( 1,0)     not null  ,
-        level_name        varchar2 (10 byte)  not null  ,
-        check_interval    number   ( 2,0)     not null  ,
+        conf_id              varchar2 (  16 byte)  not null  ,
+        conf_by              varchar2 (  64 byte)            ,
+        conf_sysdate         date                  not null  ,
+        level_id             number   (   1,0)     not null  ,
+        level_name           varchar2 (  10 byte)  not null  ,
+        check_interval       number   (   2,0)     not null  ,
+        units_level_warning  varchar2 (4000 byte)            ,
+        units_level_info     varchar2 (4000 byte)            ,
+        units_level_debug    varchar2 (4000 byte)            ,
+        units_level_trace    varchar2 (4000 byte)            ,
         --
         constraint  console_conf_pk   primary key ( conf_id                 )  ,
         constraint  console_conf_ck1  check       ( conf_id = 'GLOBAL_CONF' )  ,
@@ -71,13 +75,17 @@ begin
 end;
 /
 
-comment on table  console_conf                is 'Holds the global console configuration in a single record.';
-comment on column console_conf.conf_id        is 'The primary key - is secured by a check constraint which allows only one record in the table.';
-comment on column console_conf.conf_by        is 'The user who configured the console the last time.';
-comment on column console_conf.conf_sysdate   is 'The date when the console was configured the last time.';
-comment on column console_conf.level_id       is 'The defined global log level ID.';
-comment on column console_conf.level_name     is 'The defined log level name.';
-comment on column console_conf.check_interval is 'The number of seconds a session looks for a changed configuration.';
+comment on table  console_conf                     is 'Holds the global console configuration in a single record.';
+comment on column console_conf.conf_id             is 'The primary key - is secured by a check constraint which allows only one record in the table.';
+comment on column console_conf.conf_by             is 'The user who configured the console the last time.';
+comment on column console_conf.conf_sysdate        is 'The date when the console was configured the last time.';
+comment on column console_conf.level_id            is 'The defined global log level ID.';
+comment on column console_conf.level_name          is 'The defined log level name.';
+comment on column console_conf.check_interval      is 'The number of seconds a session looks for a changed configuration.';
+comment on column console_conf.units_level_warning is 'A comma separated list of units configured for level warning.';
+comment on column console_conf.units_level_info    is 'A comma separated list of units configured for level info.';
+comment on column console_conf.units_level_debug   is 'A comma separated list of units configured for level debug.';
+comment on column console_conf.units_level_trace   is 'A comma separated list of units configured for level trace.';
 
 
 
@@ -238,7 +246,7 @@ prompt - Package CONSOLE (spec)
 create or replace package console authid definer is
 
 c_name    constant varchar2 ( 30 byte ) := 'Oracle Instrumentation Console'       ;
-c_version constant varchar2 ( 20 byte ) := '1.0-beta3'                            ;
+c_version constant varchar2 ( 20 byte ) := '1.0-beta4'                            ;
 c_url     constant varchar2 ( 40 byte ) := 'https://github.com/ogobrecht/console' ;
 c_license constant varchar2 (  5 byte ) := 'MIT'                                  ;
 c_author  constant varchar2 ( 15 byte ) := 'Ottmar Gobrecht'                      ;
@@ -452,7 +460,7 @@ Call Stack
 {{#}}# Call Stack
 
 - PLAYGROUND_DATA.SOME_API.DO_STUFF, line 38
-- anonymous_block, line 2
+- __anonymous_block, line 2
 
 {{#}}# Error Stack
 
@@ -1143,8 +1151,12 @@ $end
 --------------------------------------------------------------------------------
 
 procedure conf (
-  p_level           integer default c_level_error, -- Level 1 (error), 2 (warning), 3 (info), 4 (debug) or 5 (trace).
-  p_check_interval  integer default 10             -- The number of seconds a session looks for a changed configuration. Allowed values: 1 to 60 seconds.
+  p_level               integer default c_level_error , -- Level 1 (error), 2 (warning), 3 (info), 4 (debug) or 5 (trace).
+  p_check_interval      integer default 10            , -- The number of seconds a session looks for a changed configuration. Allowed values: 1 to 60 seconds.
+  p_units_level_warning varchar2 default null         , -- A comma separated list of units names which should have log level warning. Example: p_units_level_warning => 'OWNER.UNIT,SCHEMA2.PACKAGE3'
+  p_units_level_info    varchar2 default null         , -- Same as p_units_level_warning for level info.
+  p_units_level_debug   varchar2 default null         , -- Same as p_units_level_warning for level debug.
+  p_units_level_trace   varchar2 default null           -- Same as p_units_level_warning for level trace.
 );
 /**
 
@@ -1569,7 +1581,7 @@ between 0 and 4.
 
 --------------------------------------------------------------------------------
 
-function  get_scope return varchar2;
+function get_scope return varchar2;
 /**
 
 Get the current scope (method, line number) from the call stack.
@@ -1581,7 +1593,19 @@ log entry.
 
 --------------------------------------------------------------------------------
 
-function  get_call_stack return varchar2;
+function get_calling_unit return varchar2;
+/**
+
+Get the calling unit (OWNER.UNIT) from the call stack.
+
+Is used internally by console to check if unit is configured for current log
+level.
+
+**/
+
+--------------------------------------------------------------------------------
+
+function get_call_stack return varchar2;
 /**
 
 Get the current call stack (and error stack/backtrace, if available).
@@ -1886,8 +1910,6 @@ c_html_less_then       constant varchar2 ( 4 byte) := chr(38) || 'lt;';
 c_html_greater_then    constant varchar2 ( 4 byte) := chr(38) || 'gt;';
 c_timestamp_format     constant varchar2 (25 byte) := 'yyyy-mm-dd hh24:mi:ss.ff6';
 c_default_label        constant varchar2 ( 7 byte) := 'Default';
-c_anon_block_ora       constant varchar2 (17 byte) := '__anonymous_block';
-c_anonymous_block      constant varchar2 (15 byte) := 'anonymous_block';
 c_conf_id              constant varchar2 (11 byte) := 'GLOBAL_CONF';
 c_client_id_prefix     constant varchar2 ( 6 byte) := '{o,o} ';
 c_console_owner        constant varchar2 (30 byte) := user;
@@ -1943,6 +1965,17 @@ c_interval_day_to_second constant pls_integer := 183;
 c_ref                    constant pls_integer := 111;
 c_ref_cursor             constant pls_integer := 102; -- same identfiers for strong and weak ref cursor
 
+type timers_tab      is table of timestamp   index by t_vc128;
+type counters_tab    is table of pls_integer index by t_vc128;
+type saved_stack_tab is table of t_vc1k      index by binary_integer;
+type unit_list_tab   is table of t_vc4k      index by binary_integer;
+
+g_timers         timers_tab;
+g_counters       counters_tab;
+g_log_cache      tab_logs := new tab_logs();
+g_saved_stack    saved_stack_tab;
+g_prev_error_msg t_vc1k;
+
 g_conf_check_sysdate        date;
 g_conf_exit_sysdate         date;
 g_conf_context_is_available boolean;
@@ -1955,16 +1988,7 @@ g_conf_user_env             boolean;
 g_conf_apex_env             boolean;
 g_conf_cgi_env              boolean;
 g_conf_console_env          boolean;
-
-type tab_timers      is table of timestamp   index by t_vc128;
-type tab_counters    is table of pls_integer index by t_vc128;
-type tab_saved_stack is table of t_vc1k      index by binary_integer;
-
-g_timers         tab_timers;
-g_counters       tab_counters;
-g_log_cache      tab_logs := new tab_logs();
-g_saved_stack    tab_saved_stack;
-g_prev_error_msg t_vc1k;
+g_conf_unit_levels          unit_list_tab;
 
 -------------------------------------------------------------------------------
 -- PRIVATE HELPER METHODS (forward declarations)
@@ -2726,7 +2750,7 @@ is
       p_application_id => v_app_id,
       p_name           => p_constraint_name,
       p_language       => apex_util.get_preference('FSP_LANGUAGE_PREFERENCE'),
-      p_message_text   => 'FIXME: Create message for constraint ' || p_constraint_name);
+      p_message_text   => 'DEVELOPER TODO: Change this message for constraint ' || p_constraint_name);
     commit;
   end;
   --
@@ -2944,11 +2968,60 @@ $end
 --------------------------------------------------------------------------------
 
 procedure conf (
-  p_level           integer default c_level_error,
-  p_check_interval  integer default 10
-)
+  p_level               integer  default c_level_error ,
+  p_check_interval      integer  default 10            ,
+  p_units_level_warning varchar2 default null          ,
+  p_units_level_info    varchar2 default null          ,
+  p_units_level_debug   varchar2 default null          ,
+  p_units_level_trace   varchar2 default null          )
 is
-  v_row console_conf%rowtype;
+  pragma autonomous_transaction;
+  v_old_conf console_conf%rowtype;
+  v_conf     console_conf%rowtype;
+  v_sep      varchar2(1 byte) := ',';
+  --
+  procedure add_unit_to_level (
+    p_unit  varchar2    ,
+    p_level pls_integer )
+  is
+  begin
+    if p_level >= 2 then v_conf.units_level_warning := v_sep || p_unit; end if;
+    if p_level >= 3 then v_conf.units_level_info    := v_sep || p_unit; end if;
+    if p_level >= 4 then v_conf.units_level_debug   := v_sep || p_unit; end if;
+    if p_level >= 5 then v_conf.units_level_trace   := v_sep || p_unit; end if;
+  end add_unit_to_level;
+  --
+  procedure close_unit_levels is
+  begin
+    if v_conf.units_level_warning is not null then v_conf.units_level_warning := v_conf.units_level_warning || v_sep; end if;
+    if v_conf.units_level_info    is not null then v_conf.units_level_info    := v_conf.units_level_info    || v_sep; end if;
+    if v_conf.units_level_debug   is not null then v_conf.units_level_debug   := v_conf.units_level_debug   || v_sep; end if;
+    if v_conf.units_level_trace   is not null then v_conf.units_level_trace   := v_conf.units_level_trace   || v_sep; end if;
+  end close_unit_levels;
+  --
+  procedure normalize_units_and_levels (
+    p_units varchar2    ,
+    p_level pls_integer )
+  is
+    v_units t_vc32k;
+    v_unit  t_vc32k;
+    v_index pls_integer;
+  begin
+    if p_units is not null then
+      v_units := p_units;
+      loop
+        v_index := instr(v_units, v_sep);
+        if v_index > 0 then
+          add_unit_to_level( trim(substr(v_units, 1, v_index - 1)), p_level );
+          v_units := substr(v_units, v_index + 1);
+        else
+          add_unit_to_level( trim(v_units), p_level );
+          exit;
+        end if;
+      end loop;
+    end if;
+  end normalize_units_and_levels;
+  --
 begin
   assert (
     p_level in (1, 2, 3),
@@ -2961,21 +3034,32 @@ begin
     p_check_interval between 10 and 60,
     'Check interval needs to be between 10 and 60 (seconds). ' ||
     'Values between 1 and 10 seconds can only be set per session with the procedure init.');
-  v_row.conf_id        := c_conf_id;
-  v_row.conf_by        := substrb(
+  v_conf.conf_id        := c_conf_id;
+  v_conf.conf_by        := substrb(
                             coalesce(sys_context('USERENV','OS_USER'), sys_context('USERENV','SESSION_USER')),
                             1,
                             64);
-  v_row.conf_sysdate   := sysdate;
-  v_row.level_id       := p_level;
-  v_row.level_name     := get_level_name(p_level);
-  v_row.check_interval := p_check_interval;
+  v_conf.conf_sysdate   := sysdate;
+  v_conf.level_id       := p_level;
+  v_conf.level_name     := get_level_name(p_level);
+  v_conf.check_interval := p_check_interval;
+  normalize_units_and_levels (p_units_level_warning, 2);
+  normalize_units_and_levels (p_units_level_info   , 3);
+  normalize_units_and_levels (p_units_level_debug  , 4);
+  normalize_units_and_levels (p_units_level_trace  , 5);
+  close_unit_levels;
   --
-  update console_conf set row = v_row where conf_id = c_conf_id;
+  v_old_conf := utl_read_global_conf;
+  --
+  update console_conf set row = v_conf where conf_id = c_conf_id;
   if sql%rowcount = 0 then
-    insert into console_conf values v_row;
+    insert into console_conf values v_conf;
   end if;
   commit;
+  --
+  if nvl(v_old_conf.level_id, 1) != v_conf.level_id then
+    utl_ctx_clear_all;
+  end if;
 end conf;
 
 --------------------------------------------------------------------------------
@@ -3466,11 +3550,7 @@ begin
     --ignore 1, is always this function (get_scope) itself
     for i in 2 .. utl_call_stack.dynamic_depth
     loop
-      --the replace changes `__anonymous_block` to `anonymous_block`
-      v_subprogram := replace (
-        utl_call_stack.concatenate_subprogram( utl_call_stack.subprogram(i) ),
-        c_anon_block_ora,
-        c_anonymous_block);
+      v_subprogram := utl_call_stack.concatenate_subprogram( utl_call_stack.subprogram(i) );
       --exclude console package from the scope
       if instr ( upper(v_subprogram), 'CONSOLE.' ) = 0 then
         v_return := v_return
@@ -3482,6 +3562,29 @@ begin
   end if;
   return v_return;
 end get_scope;
+
+--------------------------------------------------------------------------------
+
+function get_calling_unit return varchar2 is
+  v_return     t_vc32k;
+  v_subprogram t_vc32k;
+begin
+  if utl_call_stack.dynamic_depth > 0 then
+    --ignore 1, is always this function (get_scope) itself
+    for i in 2 .. utl_call_stack.dynamic_depth
+    loop
+      v_subprogram := utl_call_stack.concatenate_subprogram( utl_call_stack.subprogram(i) );
+      --exclude console package
+      if instr ( upper(v_subprogram), 'CONSOLE.' ) = 0 then
+        v_return := v_return
+          || case when utl_call_stack.owner(i) is not null then utl_call_stack.owner(i) || '.' end
+          || substr(v_subprogram, 1, instr(v_subprogram,'.') - 1 );
+      end if;
+      exit when v_return is not null;
+    end loop;
+  end if;
+  return v_return;
+end get_calling_unit;
 
 --------------------------------------------------------------------------------
 
@@ -3505,11 +3608,7 @@ begin
     --ignore 1, is always this function (get_call_stack) itself
     for i in 2 .. utl_call_stack.dynamic_depth
     loop
-      --the replace changes `__anonymous_block` to `anonymous_block`
-      v_subprogram := replace(
-        utl_call_stack.concatenate_subprogram ( utl_call_stack.subprogram(i) ),
-        c_anon_block_ora,
-        c_anonymous_block);
+      v_subprogram := utl_call_stack.concatenate_subprogram ( utl_call_stack.subprogram(i) );
       --exclude console package from the call stack
       if instr( upper(v_subprogram), 'CONSOLE.' ) = 0 then
         v_return := v_return
@@ -3539,7 +3638,7 @@ begin
     loop
       v_return := v_return
         || '- '
-        || coalesce( utl_call_stack.backtrace_unit(i), c_anonymous_block )
+        || coalesce( utl_call_stack.backtrace_unit(i), '__anonymous_block' )
         || ', line ' || utl_call_stack.backtrace_line(i) || c_lf;
     end loop;
     v_return := v_return || c_lf;
@@ -3663,6 +3762,10 @@ begin
   append_row('g_conf_apex_env',                   to_yn( g_conf_apex_env                               ) );
   append_row('g_conf_cgi_env',                    to_yn( g_conf_cgi_env                                ) );
   append_row('g_conf_console_env',                to_yn( g_conf_console_env                            ) );
+  append_row('g_conf_unit_levels(2)',                    g_conf_unit_levels(2)                           );
+  append_row('g_conf_unit_levels(3)',                    g_conf_unit_levels(3)                           );
+  append_row('g_conf_unit_levels(4)',                    g_conf_unit_levels(4)                           );
+  append_row('g_conf_unit_levels(5)',                    g_conf_unit_levels(5)                           );
   append_row('g_counters.count',                to_char( g_counters.count                              ) );
   append_row('g_timers.count',                  to_char( g_timers.count                                ) );
   append_row('g_log_cache.count',               to_char( g_log_cache.count                             ) );
@@ -4104,7 +4207,12 @@ begin
   if g_conf_check_sysdate < sysdate then
     utl_load_session_configuration;
   end if;
-  return g_conf_level >= p_level or sqlcode != 0;
+  return
+    g_conf_level >= p_level
+    or
+    sqlcode != 0
+    or
+    g_conf_unit_levels(p_level) is not null and instr(g_conf_unit_levels(p_level), ','||get_calling_unit||',') > 0;
 end utl_logging_is_enabled;
 
 --------------------------------------------------------------------------------
@@ -4218,11 +4326,10 @@ end utl_ctx_clear;
 
 procedure utl_load_session_configuration is
   v_session_conf console_sessions%rowtype;
+  v_global_conf  console_conf%rowtype;
   --
   procedure set_default_config is
-    v_global_conf  console_conf%rowtype;
   begin
-    v_global_conf := utl_read_global_conf;
     --We have no real conf until now, so we fake 24 hours.
     --Conf will be re-evaluated at least every 10 seconds.
     g_conf_exit_sysdate   := sysdate + 1;
@@ -4261,10 +4368,12 @@ procedure utl_load_session_configuration is
   end load_config_from_table_row;
   --
 begin
-  --
-
+  v_global_conf := utl_read_global_conf;
+  g_conf_unit_levels(2) := v_global_conf.units_level_warning ;
+  g_conf_unit_levels(3) := v_global_conf.units_level_info    ;
+  g_conf_unit_levels(4) := v_global_conf.units_level_debug   ;
+  g_conf_unit_levels(5) := v_global_conf.units_level_trace   ;
   if g_conf_context_is_available then
-
     g_conf_exit_sysdate := to_date(sys_context(c_ctx_namespace, c_ctx_exit_sysdate), c_ctx_date_format);
     if g_conf_exit_sysdate is null then
       set_default_config;
@@ -4274,9 +4383,7 @@ begin
     else
       load_config_from_context;
     end if;
-
   else
-
     v_session_conf := utl_read_session_conf(g_conf_client_identifier);
     g_conf_exit_sysdate := v_session_conf.exit_sysdate;
     if g_conf_exit_sysdate is null or g_conf_exit_sysdate < sysdate then
@@ -4284,9 +4391,7 @@ begin
     else
       load_config_from_table_row;
     end if;
-
   end if;
-
   g_conf_check_sysdate := least(g_conf_exit_sysdate, sysdate + 1/24/60/60 * g_conf_check_interval);
 
 end utl_load_session_configuration;