+++ docs/changelog.md
@@ -1,9 +1,15 @@
 # Changelog
 
+## v1.0-beta4 (2021-04-25)
+
+- Add the ability to set specific PL/SQL units (packages, functions, procedures)
+  to a different log level, if you want to monitor for example new code in all
+  sessions (see [console.conf](package-console.md#procedure-conf))
+
 ## v1.0-beta3 (2021-04-12)
 
 - New global configuration to be able to set the default log level and the conf
-  check interval to other values then 1 and 10 seconds (see
+  check interval to other values than 1 and 10 seconds (see
   [console.conf](package-console.md#procedure-conf))
 - Fixed: Saved call stack shows line numbers from the method console.assert
   instead of the calling code